+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 1. [14497](https://github.com/influxdata/influxdb/pull/14497): Update the documentation link for Telegraf.
 1. [14492](https://github.com/influxdata/influxdb/pull/14492): Fix to surface errors properly as task notifications on create.
 1. [14569](https://github.com/influxdata/influxdb/pull/14569): Fix limiting of get runs for task.
+1. [14779](https://github.com/influxdata/influxdb/pull/14779): Refactor tasks coordinator.
 
 ## v2.0.0-alpha.16 [2019-07-25]